+++ crates/mimic_build/src/actor/query.rs
@@ -47,7 +47,7 @@ fn generate_query(name: &str, builder: &ActorBuilder, kind: QueryKind) -> TokenS
                     #entity_path => db!().delete().response::<#ty>(query)
                 },
                 QueryKind::Save => quote! {
-                    #entity_path =>  db!().save().execute::<#ty>(query)
+                    #entity_path =>  db!().save().response::<#ty>(query)
                 },
             }
         });
@@ -71,7 +71,7 @@ fn generate_query(name: &str, builder: &ActorBuilder, kind: QueryKind) -> TokenS
             pub fn #fn_name(
                 path: String,
                 query: ::mimic::db::query::LoadQuery,
-            ) -> Result<::mimic::db::response::LoadResponse, ::mimic::Error>
+            ) -> Result<Vec<::mimic::core::Key>, ::mimic::Error>
         },
 
         QueryKind::Save => quote! {
@@ -87,7 +87,7 @@ fn generate_query(name: &str, builder: &ActorBuilder, kind: QueryKind) -> TokenS
             pub fn #fn_name(
                 path: String,
                 query: ::mimic::db::query::DeleteQuery,
-            ) -> Result<::mimic::db::response::DeleteResponse, ::mimic::Error>
+            ) -> Result<Vec<::mimic::core::Key>, ::mimic::Error>
         },
     };